requirements.txt
flask
requests
beautifulsoup4
dnspython
python-whois
tldextract
fuzzywuzzy
matplotlib
python-Levenshtein